Updated README for Clarity on Params
The readme can be clearer to present the fact that there are parameters for specifying continuous variables as we only knew about it after diving into the actual code. The `continuous_cols` is not resented in Optional Arguments either. I hope this adds clarity for your users and readers. Thanks!